nexum-backend/ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

# GCP
google-cloud-storage==2.18.2
google-cloud-pubsub==2.23.0

# Utils
python-multipart==0.0.9
httpx==0.27.2
structlog==24.4.0
